We're back to patch 17 now. Checkpoint. Move tapes/root from root211 -> root-2.11 Fix a bug in patch extraction Optmize a bit more Extract /dev/[A-Z] from the root image
To be fair, there's 1 missing file (KAZOO config) and 3 files that
don't patch quite right that we need to sort out. ld.c is messed up,
but patches back to 1 now. We may need to create a 2.11 pl -1 that
has it form 2.10.1 so we can regenreate a system.

This is a good checkoint since 13-17 are all dealing with pcc/lint
and other ugliness and this is the first instance of rm/cp/tar xvf
in a little while.
plus we need to write the tar patch code for 2.tar and 3.tar
(which are really just multiple parts in 1 instead of over several
patch levels, but are one patch per file where later parts
often combine them).