Work a bit on ld tonight
It looks like 2.11BSD didn't use new ranlib structures, but still used the
2.10.1 fixed-length table. So use that code for case 2: of load1arg(). The code
that I'd stolen from 4.3BSD assumes ranlib() header, not the pdp-11 tab. There's
also several globals it touched for the 'new format' .o files, which is akin to
the stuff done around patch 160 to get to long filenames, but done in a more
extensive way.

It also looks load1() shouldn't have been copied from the VAX version. First,
mget() used in the vax version has an extra arg, so that can't be right. Second,
the reloc stuff is absent from the pdp-11 version. so go with the 2.10.1 version
instead.

These two fixes get me a lot further in tracking now the putative ld.c.

There's 29 extra lines still, which need to be investigated.
...
Hunk #16 succeeded at 550.
Hunk #17 succeeded at 900 (offset 29 lines).
Hunk #18 succeeded at 1102 (offset 29 lines).
...

Now, the fix that was posted in comp.bugs.2bsd might be the bulk of it:
 ld.c |   26 +++++++++++++++++++++++!!!
 1 file changed, 23 insertions(+), 3 modifications(!)
but that starts at line 1125, which may be a bit late...  this needs to
be investigated. But I think that I found those modifications because
I hit a snag earlier in this procress. It's too late to track those down.