ranlib.c most likely was 2.10.1 with some 4.3BSD stuff thrown in rather than a full copy of 4.3BSD
Why? 4.3BSD's produced the full ranlib table, while we're now pretty sure that
2.11BSD's ld.c produced the old format from 2.10.1, but in the new portable
archive. The 2.10.1 version does that, but didn't support the portable archive
format. It's clear from code examination that both these ranlib.c files trace to
a common ancestor. It would be super easy to copy bits and pieces from 4.3BSD
without doing the whole thing.

So, introduce a patch which replaces a few lines in main that validates the
archive as the proper type and the entire nextel, fixsize and fixdate functions.