Make MAKEDEV work.
It looks like there was a copy of MAKEDEV hiding in usr/src/sys/pdpdist/MAKEDEV
this whole time. dev/MAKEDEV was patched, but this one was not. Therefore,
it should be what we want in /dev/MAKEDEV. In theory this is correct, but in
practice, we'll see. It's a lot closer, though, than anything else.

Note to self: when we roll forward, each patch that touches /dev/MAKEDEV
should copy to usr/src/sys/pdpdist/MAKEDEV as well.