Revamp how we bootstrap as
OK. Use v7's assembler to assemble the 2.11 system call .s files. We have to use
the 2.11 cpp, though, because they depend on syscalls.h which has long #define
names. We create a dummy signal. signal is used to ignore and/or cleanup after
SIGINT. Turns out, that doesn't matter, it seems.  This creates a bootstrap.o
that we use to build as and as2. It's a little inefficient, since it adds a few
extra bytes of text since there's a couple of system calls only used by
one. However, the signal implementation more than makes up for it, so it should
be good.

This is untested, though.