Tweaks to the script to make as and as2 bootstrap work
v7's (and presumably 2.11's) ld -r will have a non-zero exit value. This stopped
the script running with no error message. The ld actually succeeds. Follow the
lead of v7/2.11BSD and ignore it as well.

Also, confirmed that as and as2 work on the 2.11BSD pl 195 kernel to assemmble
the pl0 assembler sources.

Next up: working on the bootstrapping script that will walk through the
dependencies one at a time to build the tools to build everything else...