Fix ranlib to actually build.
The 4.3 ranlib had a tssiz global variable, but the old code didn't. So checking
and adjusting it in fixsize() was bogus.
Likewise, we have a global variable offdelta, so use that instead of the one that
got added to fixsize().
The symbol stuff is simpler too, since we don't have the paging code here, so match
2.10.1BSD's ranlib.c more closely here.

With these changes, it almost builds on 2.11BSD pl 195, but fails to link. We'll
revisit this once we have the rest of the stuff bootsrapped. It looks like all
we need are as, ld and nm (though ar might be needed: the ar format is the same
between 2.11BSD pl 195 and pl 0, but the program itself is completely different).