In experiments with bootstrapping, there were a number of failures with time and/or permissions (the error messages aren't entirely helpful). I had to delete the bootstrap .o's for the build to work (well, and to fix a number of bugs), so do that here.
I also need to start a script that builds ld, copies as, as2, then does the
chroot and starts to bootstrap.  Oh, wait, that failed because one needs to
build libc, which uses the old-school `lorder *.o | tsort` which uses nm too.
The list is starting to get long. ranlib didn't compile, but things worked with
/bin/true was copied to ranlib, at least to bootstrap things enough I think...