Turns out we need to do more in compat-4.1: 1. create a SYS.h symlink to ../sys/SYS.h 2. sed -e s/DEFS.h/SYS.h/h on the Makefile and reset.s file 3. Create a profiled directory
DEFS.h was renamed to SYS.h between 2.10.1BSD and 2.11BSD.

We likely need to bump the Makefile version as well, but I've
not done that since I've not tried to 'guess' the proper SCCS
ID elsewhere. Without a good notion of what date to use, I'd
be shooting in the dark anyway.