OK. Fall back to the 2.10.1 ranlib.c
But fix it to work with new-style archives.
Need to declare atol(), keep track of the size of the archive
and space out the archive name. The big missing thing from the
last version was remembering that you have to declare anything
that doesn't return int, and int != long in pdp11 land...

We almost work with this... I think there's a bug in 'ar' because
when we replace something we get garbage after __.SYMDEF name.