Need to fix fixdate too.
Need to apply the same fixes to fixdate that we did to other locations now that
the fields are printable. This fixes the weird archive member name that we're
seeing. This keeps us from corrupting the __.SYMDEF name after ar had
successfully inserted it into the library.

Note: we're not working... ld doesn't like the added table of contents so
something is still amiss.