Three main changes:
First, copy the precious bootstrap binaries (ld, nm, as and as2) to /bs
in build.sh so that build2.sh can be run multiple times.

Second, we need to build and install install(1) before we use install -s
as part of 'make install'

Third, we have numeric libraries, so include that in the expanding pattern.