Cope with the fact that libom.a provides -lm functionality.
The usr.lib/Makefile install target does this link. Since we don't use that
target directly for bootstrapping for other reasons, we need to do it here.

There's a couple of other .o's that get installed in a weird way. I don't
think we need them to bootstrap, but something to keep in mind.