Commit the ranlib -t patch from comp.bugs.2bsd
I'd do this directly from the comp.bugs.2bsd archive, but that conflicts with
the subsequent work I've recreated. So I have to apply it by hand. I might be
able to do it directly and regen the diff, but for now capture this.

This should fix the errors with libutil.a and ranlib -t that i'm still seeing in
the reconstruction...