Add the reconstructed tmscpboot.s file. We have evidence from the 2.11BSD pl2 and newer mtboot.s that the tmscp support came from the ultrix-11 boot loader. This bootloader is in the 2.11 tree as sys/OTHER/tk/pdpstand/tkboot.s. It appears the SMS ported this to 2.11BSD in late 1990, and then combined it with mtboot.s in mid 1991 (for patch 2).
Based on that, I made minimal changes to tkboot.s after copying it to
tmscpboot.s. These changes removed bogus clearing of the core below the boot
loader, added clearing of the area above (used for the tmscp packet buffers) and
removing a bogus test for the 407 magic number (which can never happen). These
are consistent with the start of mtboot.s, and would likely have been the
starting point for SMS' work in this area. I've not yet tested to see if this
will work, but tracing through the code suggets to me there's a high likelihood
that it will.

We also need to create a tmscpboot.data file. I'm pretty sure this is the same
as the normal maketape.data file with this loader instead of the default one.

Note: this will mean I need 2 images; one for Mag Tape images (TS, HT, etc) and
one for TK-50 (the TMSCP class).