OK. At patch 80 (really 78), a new tape was cut. This tape had all the fixes rolled up into one tape. However, it added cxref, mtools, zmodem, and tcsh to the mix. These all need to be removed.
And sometime, unknown, Kermit 5A(188) was added to the tree. It was released 23
Nov 92, but the rollup patches were done about a month earlier.  There's no good
place to drop it in, though, in the timeline so we're pretending it happened
here. To undo it, we're removing it and adding back the kermit 4E(70) from
2.10.1BSD. There was a slightly newer version released after 2.10.1, but before
2.11BSD. I've opted not to update to that, since it didn't offer anyhing
special. It falls around patch 86 or so in the progress if it was released the
same day. We know it was sometime before patch 106-116 fixups were done in Jan
93.  I'll reconstruct it as just after patch 80 in the git repo with no real
patch number.

All this came about because /usr/src/new and /usr/src/local were included in the
2.11BSD, but had at-best quasi-official status. So they were updated, but no
patches were rotuinely generated until after the patch 195 tape. So, to try to
reconstruct the 2.11BSD tape, we have to make our best guesses here. And some of
those are that there was no version chasing of every version released, while the
second was that if it was in 2.10.1BSD, it was in 2.11BSD as the same release
(unless we have evidence to the contrary). This allows us to recover kermit from
2.10.1BSD w/o chasing the slightly newer version.