Recover a patch to csh that wasn't published in the original patch train, but that did show up in the patch-80 catchup mega patch.
Examining the dates, I noticed that the csh dates were wrong. It turns out that
there was a discrepency. When I went looking, I found a patch in xac of the
patch-80 catch-up patches that matches up perfectly. Apply it as patch 18a,
since that's when it falls in the dates...