gen211git: script to create 2.11BSD repo
This is the start of gen211git. This script takes the reconstructed system image
in root-2.11, extracts the sources, and then checks them into a freshly created
git repo.

In the future, it will apply all 470-odd patches, one at a time, and commit them
to that repo with a reasonable commit message. In addition, it commits the
seemlying missing patches so that we match at each of the 'way points' we have
preserved: p195, p277, p431, and p457.

It will use some of the same helper scripts that mk211bsd used to walk backwards,
as well as new scripts to help it walk forward (since often mk211bsd too some
short-cuts knowing it wouldn't matter to the end goal)....